smart-learning-backend/python-services/requirements.txt
Flask==3.0.3
tensorflow==2.17.0
Pillow==10.3.0
SpeechRecognition==3.10.1
numpy==1.26.4
google-generativeai==0.4.1
flask-cors==4.0.1
pydub==0.25.1
python-dotenv==1.0.1
librosa==0.10.2